app/requirements.txt
altgraph==0.17.2
async-generator==1.10
attrs==21.2.0
auto-py-to-exe==2.13.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.0
Eel==0.12.4
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.12.0
humanize==3.13.1
hurry.filesize==0.9
idna==3.3
outcome==1.1.0
pefile==2021.9.3
Pillow==8.4.0
pycparser==2.21
pycryptodome==3.12.0
pycryptodomex==3.12.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pyodbc==4.0.32
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pywin32-ctypes==0.2.0
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
webdriver-manager==3.5.2
whichcraft==0.6.1
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.4.0

mariadb
pymongo
pymysql
mysql-connector-python
redis
schedule
python-dotenv